mi_backend_python/requirements.txt
fastapi
pydantic
python-dotenv
httpx
pandas
uvicorn
python-multipart
